Optimize the Pokemon Hunter a bit more
Will notice when it gets stuck on the map for some reason. This
sometimes happen, I don't know why, but this will stop the bot from
standing in place while it's supposed to be 'Hunting'.
This also adds the reason for a hunt. And it will catch VIP and unseen
Pokemons while "Locked on".