Pokemon Hunter - Hunting trash
The hunter will check the space left in the Pokemon bag, when it is
below 20 slots, the hunter will start hunting 'trash' Pokemon. Only
Caterpie, Weedle and Pidgey (and their evolutions) are hunted down.
These Pokemon require only 12 candy to evolve.

No worries about not hunting VIP Pokemon (and their family if set)
because the hunter will always check if there is an VIP Pokemon nearby,
if so then that will take priority.