Daily catch limit check added
This adds a check for the daily catch limit. It will disable all
catching tasks (just as we're out of balls) for an hour. After that it
will check if the daily catching amount is okay again.
If the daily catch limit passes by at least 20 during that hour, the
catching tasks are also resumed.