Relure cluster function
When camping forts and we have lures to spare, when a lure ends we will
put a lure in the Pokestop that has no lure. This will also lure unlured
forts if we arrive at a cluster with a missing lure. You can configure
how many lures you want to keep (default 1)

Before we relure a Pokestop, we check if we will sleep before the lure
will end. If so, no lure will be deployed.